fixes #88, #113
