+++ pymer4/tests/test_bridge.py
@@ -1,4 +1,4 @@
-from pymer4.bridge import pandas2R, R2pandas, numpy2R, R2numpy
+from pymer4.bridge import pandas2R, R2pandas, numpy2R, R2numpy, con2R
 from rpy2 import robjects as ro
 import numpy as np
 import pandas as pd
@@ -49,3 +49,19 @@ def test_numpy():
     letters_fromR_toR = numpy2R(letters_fromR)
     # Even we we convert back rpy2 keeps it as an array
     assert not isinstance(letters_fromR_toR, ro.vectors.StrVector)
+
+
+def test_con2R():
+    x = np.array([[-1, 0, 0, 1], [-0.5, -0.5, 0.5, 0.5], [-3 / 3, 1 / 3, 1 / 3, 1 / 3]])
+    out = con2R(x)
+    assert out.shape == (4, 3)
+    names = ["1 v s4", "1+2 vs 3+4", "1 vs 2+3+4"]
+    out = con2R(x, names=names)
+    assert isinstance(out, pd.DataFrame)
+    assert [x == y for x, y in zip(out.columns, names)]
+    assert out.shape == (4, 3)
+
+    out = con2R(np.array([-1, 0, 1]))
+    assert np.allclose(
+        out, np.array([[-0.5, 0.40824829], [0.0, -0.81649658], [0.5, 0.40824829]])
+    )
